requirements.txt
fastapi~=0.128.0
httpx~=0.28.1
crewai[tools]
litellm>=1.81.6
litellm[proxy]
python-dotenv